test(prosecute): split coverage into per-concern files (closes #68)
packages/prosecute had a single 673-line test/prosecute.test.mjs covering
435 lib lines (lib/run.mjs + lib/schema.mjs) -- the P5 evidence ledger's
dry-lens/consecutive-dry/revision-binding trust-boundary logic. Split it
into test/helpers.mjs (shared fixtures) plus six per-concern files mirroring
packages/gate-fuzzing's pattern:

- prosecute-dry-lenses.test.mjs: finalDryLenses()/dryLenses.size>=3 boundary
  (2 distinct trailing lenses fails, 3 passes, repeats don't double-count)
- prosecute-consecutive-dry.test.mjs: consecutiveDry increment/reset,
  asserted per-pass, including that a non-dry pass resets the streak
- prosecute-revision-binding.test.mjs: resolveRevision/transcriptProof/
  reviewPacketProof/seedOpenFindingsFromManifest, with new explicit
  staleness-rejection cases and a stale-revision-doesn't-block-fresh-
  revision case
- prosecute-findings-lifecycle.test.mjs: open/killed finding identity
  tracking within a run
- prosecute-schema-validation.test.mjs: op-error paths
- prosecute-cli.test.mjs: bin/adlc-prosecute.mjs subprocess smoke tests

Test-only change; no lib/run.mjs or lib/schema.mjs behavior modified. No
behavioral bug was found while writing the new boundary/staleness cases --
one test-authoring mistake (a fixture file-path collision) was fixed in the
test itself, not the library. Added docs/specs/prosecute-coverage-split.md
as a lightweight P1 record; verified with adlc spec-lint.

Verified: packages/prosecute npm test (32/32 pass) and the full per-package
test loop across the monorepo (no regressions).
